+++ CHANGELOG.md
@@ -134,6 +134,9 @@
     * Renamed to `from_xy_array()`.
     * Is now a `classmethod` instead of `staticmethod`.
     * Refactored to make code simpler.
+* `KeypointsOnImage.get_coords_array()`
+    * Marked as deprecated.
+    * Renamed to `to_xy_array()`.
 
 
 # 0.2.8